handle delete user on cloudinary

+++ src/user/user.service.ts
@@ -85,15 +85,49 @@ export class UserService {
 
   async deleteUser(userId: string) {
     try {
-      const deletedProfile = await this.prismaService.profile.delete({
-        where: {
-          userId,
-        },
+      let deletedProfile = null;
+      let deletedEvent = null;
+      let deletedMessages = null;
+
+      const profile = await this.prismaService.profile.findUnique({
+        where: { userId },
       });
+
+      if (profile) {
+        deletedProfile = await this.prismaService.profile.delete({
+          where: {
+            userId,
+          },
+        });
+      }
+
+      const events = await this.prismaService.event.findMany({
+        where: { creatorId: userId },
+      });
+      if (events.length > 0) {
+        for (const event of events) {
+          const messages = await this.prismaService.message.findMany({
+            where: {
+              eventId: event.id,
+            },
+          });
+          if (messages.length > 0) {
+            const deletedMessage = await this.prismaService.message.deleteMany({
+              where: {
+                eventId: event.id,
+              },
+            });
+            deletedMessages.push(deletedMessage);
+          }
+        }
+        deletedEvent = await this.prismaService.event.deleteMany({
+          where: { creatorId: userId },
+        });
+      }
 
       const deletedFolder = await this.cloudinaryService.deleteFolders(userId);
 
-      const user = await this.prismaService.user.delete({
+      const deletedUser = await this.prismaService.user.delete({
         where: {
           id: userId,
         },
@@ -105,7 +139,13 @@ export class UserService {
           email: true,
         },
       });
-      return { ...user, ...deletedProfile, ...deletedFolder };
+      return {
+        ...deletedUser,
+        ...deletedProfile,
+        ...deletedEvent,
+        ...deletedMessages,
+        ...deletedFolder,
+      };
     } catch (error) {
       throw new UnprocessableEntityException(error);
     }
